src/main/java/com/xsc/ApplicationConfig.java
package com.xsc;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.stereotype.Component;

/**
 * Created by XSC on 2017/5/5.
 */
@ConfigurationProperties(prefix = "study")
@Component
public class ApplicationConfig {

    private String name;
    private Integer age;

    public String getName() {
        return name;
    }

    public Integer getAge() {
        return age;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setAge(Integer age) {
        this.age = age;
    }
}
